core_plugin: provide fallback tmp directory
Provides manual generaton and verification of
the tmp directory path. On some UNIX systems
I found that standard c++17 method does not return
"/tmp" direcotry as it is stated in the documentation.
In those situations the method failed and returned "".

This commit provides a fallback method, that will try
to find "/tmp" directory in the filesystem after
the method std::temp_directory_path fails.

If this method fails too, user is asked to provide the TMP
directory path manually by setting environment ariable
TMPDIR.