FormatUtils: do not strip .imp from name
As .imp is added to the functions that are mostly dynamically linked
removing it leaves name of the same function that should be imported.
This leads in many cases to manually insertion of such function into
decompiled code and in the output of r2 decompilation.

Reason why I have decided to remove the .imp from name is that it is not
good for the user experience and it causes some error during testing.