r2info: leave the fcn prefix for now
This commit provides occasional returing of prefix ".fcn" to the
function.

The reason for leaving the `fcn.` prefix is that without it some
functions will have name consising only from digits [0-9a-zA-Z]+ which is not only
invalid identifier but it looks ugly too.

Example:
`void 004112e4(int a, int b)` -> `void fcn_004112e4(int a, int b)`

The other remark that should be made is that the convention in Radare2
is fcn.004112e4 (having `fcn` with dot as prefix, not underscore). Right
now is the exchange of dot to underscore made automatically in RetDec.
This may be, however, changed in the future.