core_plugin: link stdc++fs library (#12)
Provides linking of the standard C++17 filesystem library with
resulting plugin. The reason for this is such that the option to
ommit linkage of this library was introduced in GCC 9 and thus
would leave users with GCC 8 with runtime errors.

This change was tested on GCC 8 successfully.

Solves https://github.com/avast/retdec-r2plugin/issues/12