CMakeLists.txt: link stdc++17 only if found
This commit provides addition to the changes brought by commit 4219cd5.
The change is to check whether library stdc++17 exists and if so
link it to the core_plugin.

This was initiated by the fact, that not every compiler must have
stdc++17 library as it is not enforeced by standard.