r2plugin-ci: provide only building for Windows
Provides turn off of the regression testing on Windows
on Github Actions. The reason for that is that
after installation of radare2 the output of
radare2 --version is:

```
Error: R2PIPE_(IN|OUT) environment not set
Usage: r2 -c '!*r2p x' # run commands via r2pipe
```

The reason for this problem is for now unknown and
requires further investigation. After the problem is
solved this commit can be reverted.