README.md: mention updates on Radare2 and Cutter
After refactorization of Cutter cmake setting source dir of
the Cutter is no longer required. Sources of cutter are
fetched on demand.

Also there is no need to have latest Radare2 4.5.0 in order
to use the plugin. The reason is that sources of the RAnnotatedCode
were merged to the radare2 and are no longer required in this
repo.